Correccion de observaciones
Correccion de observaciones añadidas por Raul al PR
